+++ docs/api/changelog.rst
@@ -7,6 +7,7 @@ Changelog
 
 0.5.1 - TBD
     * Made FileParser do mypy style inline comments by default
+    * It is now possible to create notices for comparison using regexes and globs.
 
 .. _release-0.5.0: